Start drafting string API support and add aliases.
This changes the way colorize works and is still a *really* rough draft.
I didn't want to repeat the color codes, but also don't feel much like
wasting time making such a trivial use case non-verbose.

This relates to #1, but, as I said is still just an API idea, not
something I'd like to push to dub yet.